requirements.txt
BeautifulSoup==3.2.1
Django==1.9c1
Pillow==3.0.0
amqp==1.4.7
anyjson==0.3.3
argparse==1.2.1
backports.ssl-match-hostname==3.4.0.2
billiard==3.3.0.21
celery==3.1.19
chardet==2.3.0
distribute==0.6.24
dnspython==1.12.0
docker-py==1.5.0
ipaddr==2.1.11
ipwhois==0.11.0
kombu==3.0.29
psycopg2==2.6.1
python-Wappalyzer==0.2.2
python-magic==0.4.10
pythonwhois==2.4.3
pytz==2015.7
redis==2.10.5
requests==2.8.1
sh==1.11
six==1.10.0
tldextract==1.7.2
u-msgpack-python==2.1
websocket-client==0.34.0
wsgiref==0.1.2
